requirements.txt
python-telegram-bot
python-dotenv
requests
datetime
readability-lxml